pyproject.toml
[tool.poetry]
name = "mad_project"
version = "0.1.0"
description = ""
authors = ["rysbai <rysbai.coder@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.3"
djangorestframework = "^3.13.1"
psycopg2 = "^2.9.3"
django-database-url = "^1.0.3"
PyJWT = "^2.3.0"
factory-boy = "^3.2.1"
pytest-django = "^4.5.2"
python-dotenv = "^0.19.2"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
